3rd week
While going in-depth of the research paper published by Nvidia for Nemo we found out that they incorporated ANNOY for prompt classification and then we used it with our own chatbot and saw improved results. Moving forward with the research paper we discovered that they are leveraging the use of Rasa, Microsoft Luis and Google Dialogue flow for Natural language understanding and Dialogue Management, in which Rasa Framework is pretty flexible for the usecase and utilizes DIET classifier model. 3rd week research is on-going.